Condense tests
